findMatchingMedia > findApplicableMedia
